Refactor navigation module structure
